Added all the demos into the WebAssembly source, but not yet showing

+++ lib80GFX/lib80GFX_Main.cpp
@@ -104,7 +104,7 @@ namespace lib80GFX
 {
 	namespace MetaOut
 	{
-		void Add( MetafileWriter &metaWriter, intptr_t x, intptr_t y )
+		void Add( MetafileWriter &metaWriter, int32_t x, int32_t y )
 		{
 			metaWriter.Add( x );
 			metaWriter.Add( y );
